Uploaded Analysis 3
In this analysis, we will explore two approaches for modelling data with unusual distributions. One of these is to consider a subset of the data, narrowing the focus of our research question in order to facilitate meaningful analysis. The other is data transformation, which we have used previously (such as in taking logs of the likes variate) and will now extend to other, more complex transformation procedures.