Uploaded Analysis 4
This analysis begins with hypothesis testing, looking first at a new variate: the use of hashtags. We’ll see how often users are employing this feature, and compare that with past data to see if behaviours have changed. Understanding how user behaviour changes over time (or even if it changes at all) is of key interest to product developers (and newly-installed CEOs), and can inform whether to keep, add, or remove features. This is an opportunity for us to explore that very important principle.
We will then extend the sub-analyses of Analysis 3, where we explored how often users are tweeting. Whereas in Analysis 3 we explored this using confidence intervals, we will now deepen our understanding using hypothesis testing.
Our third sub-analysis extends our investigation into the relationship between tweet length and the number of likes received, which we began in Analysis 2. Whereas then we were limited to inspecting scatterplots and sample correlation, we are now able to employ the extremely powerful technique of linear regression!
We’ll finish off with a final sub-analysis that explores what variates affect the number of likes a tweet receives. This will build on sub-analyses conducted in Analysis 1 and 2, and you’ll get to decide which analysis you’re most interested in extending!